k8s definitions: do not use hostPort
There was three "mongo" deployments listening on the same
containerPort=27017. Binding the same port on the host is not necessary
because we use Kubernetes services to proxy the connections. It prevents
the pods to be scheduled on the same host.

Also, remove some obsolete comments referring to type=LoadBalancer that
has been removed a while ago.